requirements.txt
tsdat>=0.7.0
pydantic<2.0.0
ruamel.yaml
typer
openpyxl
pandas
